+++ source/extensions/omni.isaac.lab/docs/CHANGELOG.rst
@@ -1,3 +1,10 @@
+Changed
+^^^^^^^
+
+* Changed the data type "depth" to "distance_to_camera" in :class:`omni.isaac.lab.sensors.TiledCamera` class to be
+  consistent with all other cameras.
+
+
 Changelog
 ---------